+++ apps/evm/src/clients/api/index.ts
@@ -27,6 +27,9 @@ export * from './mutations/useSwapTokensAndRepay';
 export * from './mutations/useRepay';
 export * from './mutations/useSwapTokens';
 export * from './mutations/useWithdraw';
+export * from './mutations/useSupplyToLiquidityHub';
+export * from './mutations/useWithdrawFromLiquidityHub';
+export * from './mutations/useMigrateCoreSupplyToLiquidityHub';
 export * from './mutations/useImportSupplyPosition';
 export * from './mutations/useSetEModeGroup';
 export * from './mutations/useOpenLeveragedPosition';
@@ -274,12 +277,6 @@ export * from './queries/getProposalMinQuorumVotes/useGetProposalMinQuorumVotes'
 
 export * from './queries/getVTokenUtilizationRate';
 export * from './queries/getVTokenUtilizationRate/useGetVTokenUtilizationRate';
-
-export * from './queries/getBurnedWBnb';
-export * from './queries/getBurnedWBnb/useGetBurnedWBnb';
-
-export * from './queries/getImportablePositions';
-export * from './queries/getImportablePositions/useGetImportablePositions';
 
 export * from './queries/getBurnedWBnb';
 export * from './queries/getBurnedWBnb/useGetBurnedWBnb';