+++ apps/evm/src/components/Icon/icons/index.ts
@@ -108,3 +108,4 @@ export { default as medal } from './medal';
 export { default as trophy } from './trophy';
 export { default as swap } from './swap';
 export { default as connect } from './connect';
+export { default as hub } from './hub';